tests/wpt/tests/css/css-sizing/parsing/width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic & Extrinsic Sizing Test: parsing width with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-width">
<meta name="assert" content="width supports only the grammar 'auto | <length-percentage> | min-content | max-content | fit-content(<length-percentage>)'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("width", "none");
test_invalid_value("width", "min-content max-content");

test_invalid_value("width", "-10%");
test_invalid_value("width", "-0.5em");
</script>
</body>
</html>
